Rename models path -> controllers
